feat(header-v2): move old LayoutNavigation to LayoutNavigationLegacy
