+++ examples/magento-graphcms/components/Layout/LayoutNavigationLegacy.tsx
