+++ examples/magento-graphcms/components/Layout/index.ts
@@ -1,4 +1,4 @@
-export * from './LayoutNavigation'
+export * from './LayoutNavigationLegacy'
 export * from './LayoutMinimal'
 export * from './LayoutOverlay'
 export * from './Layout.gql'
