+++ examples/magento-open-source/components/Layout/LayoutNavigationLegacy.tsx
