+++ examples/magento-storyblok/components/Layout/LayoutNavigationLegacy.tsx
